+++ tests/functional_tests/tests_onnx_trt/test_export.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 import logging
+import os
 import shutil
 import subprocess
 import tempfile
@@ -33,6 +34,7 @@ def tmp_dir():
         logger.warning(f"Error removing temporary directory {tmp_dir}: {e}")
 
 
+@pytest.mark.skip(reason="Temporarily disabled")
 class TestONNXTRTExport:
     def test_export_onnx_trt_embedding(self):
         subprocess.run(
@@ -48,6 +50,10 @@ class TestONNXTRTExport:
                 "--normalize",
             ],
             check=True,
+            env={
+                **os.environ.copy(),
+                "HF_DATASETS_CACHE": "/tmp/hf_datasets_cache",
+            },
         )
 
     def test_export_onnx_trt_embedding_int8(self):
@@ -71,6 +77,10 @@ class TestONNXTRTExport:
                 "6",
             ],
             check=True,
+            env={
+                **os.environ.copy(),
+                "HF_DATASETS_CACHE": "/tmp/hf_datasets_cache",
+            },
         )
 
     def test_export_onnx_trt_reranking(self):
@@ -86,4 +96,8 @@ class TestONNXTRTExport:
                 "/home/TestData/llm/models/llama-3.2-nv-reranker-1b",
             ],
             check=True,
+            env={
+                **os.environ.copy(),
+                "HF_DATASETS_CACHE": "/tmp/hf_datasets_cache",
+            },
         )